MenuApp/requirements.txt
alembic==1.9.2
asyncpg==0.27.0
alembic==1.9.2
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
fastapi==0.89.1
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pydantic==1.10.4
pytest_asyncio==0.20.3
pytest==7.2.1
python-dotenv==0.21.1
redis==4.4.2
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
